+++ janitor/functions/complete.py
@@ -1,6 +1,5 @@
 from typing import Optional, Union, List, Tuple, Dict, Any
 from pandas.core.common import apply_if_callable
-import numpy as np
 import pandas_flavor as pf
 import pandas as pd
 import functools
@@ -32,14 +31,6 @@ def complete(
     If a dictionary is passed,
     the user is required to ensure that the values are unique 1-D arrays.
     The keys in a dictionary must be present in the dataframe.
-
-    `complete` can also be executed on the names of the index -
-    the index should have names
-    (`df.index.names` or `df.index.name` should not have any None).
-    Groupby is not applicable if `completing` on the index.
-    When `completing` on the index, only the passed levels are returned.
-
-
 
     Examples:
         >>> import pandas as pd
@@ -119,14 +110,14 @@ def complete(
         ...     sort=True
         ... )
            group  item_id item_name  value1  value2
-        0      1        1         a       1       4
-        1      1        2         a       0      99
-        2      1        2         b       3       6
-        3      1        3         b       0      99
-        4      2        1         a       0      99
-        5      2        2         a       0       5
-        6      2        2         b       0      99
-        7      2        3         b       4       7
+        0      1        1         a     1.0     4.0
+        1      1        2         a     0.0    99.0
+        2      1        2         b     3.0     6.0
+        3      1        3         b     0.0    99.0
+        4      2        1         a     0.0    99.0
+        5      2        2         a     0.0     5.0
+        6      2        2         b     0.0    99.0
+        7      2        3         b     4.0     7.0
 
         Limit the fill to only implicit missing values
         by setting explicit to `False`:
@@ -147,47 +138,9 @@ def complete(
         6      2        2         b     0.0    99.0
         7      2        3         b     4.0     7.0
 
-        Expose missing values via the index:
-        >>> index = pd.date_range('1/1/2000', periods=4, freq='T')
-        >>> series = pd.Series([0.0, None, 2.0, 3.0], index=index)
-        >>> DT = pd.DataFrame({'s': series})
-        >>> DT.index.name = 'dates'
-        >>> DT
-                               s
-        dates
-        2000-01-01 00:00:00  0.0
-        2000-01-01 00:01:00  NaN
-        2000-01-01 00:02:00  2.0
-        2000-01-01 00:03:00  3.0
-        >>> dates = {'dates':lambda f: pd.date_range(f.min(), f.max(), freq='30S')}
-        >>> DT.complete(dates)
-                               s
-        dates
-        2000-01-01 00:00:00  0.0
-        2000-01-01 00:00:30  NaN
-        2000-01-01 00:01:00  NaN
-        2000-01-01 00:01:30  NaN
-        2000-01-01 00:02:00  2.0
-        2000-01-01 00:02:30  NaN
-        2000-01-01 00:03:00  3.0
-
-        The above can be solved in a simpler way with `pd.DataFrame.asfreq`:
-
-        >>> DT.asfreq(freq='30S')
-                               s
-        dates
-        2000-01-01 00:00:00  0.0
-        2000-01-01 00:00:30  NaN
-        2000-01-01 00:01:00  NaN
-        2000-01-01 00:01:30  NaN
-        2000-01-01 00:02:00  2.0
-        2000-01-01 00:02:30  NaN
-        2000-01-01 00:03:00  3.0
-
     Args:
         df: A pandas DataFrame.
-        *columns: This refers to the columns to be
-            completed. It can also refer to the names of the index.
+        *columns: This refers to the columns to be completed.
             It could be column labels (string type),
             a list/tuple of column labels, or a dictionary that pairs
             column labels with new values.
@@ -233,7 +186,6 @@ def _computations_complete(
     (
         columns,
         column_checker,
-        index,
         sort,
         by,
         fill_value,
@@ -260,30 +212,23 @@ def _computations_complete(
             df=df,
             columns=columns,
             all_scalars=all_scalars,
-            index=index,
             sort=sort,
         )
     else:
         column_checker = by + column_checker
-        uniques = df.groupby(by, group_keys=True)
-        # apply is basically a for loop
-        # for scenarios where Pandas does not have
-        # a vectorized option
+        uniques = df.groupby(by, sort=False)
         uniques = {
             key: _generic_complete(
                 df=value,
                 columns=columns,
                 all_scalars=all_scalars,
-                index=index,
                 sort=sort,
             )
             for key, value in uniques
         }
-        uniques = pd.concat(uniques, names=column_checker, copy=False)
-        if not index:
-            uniques = uniques.droplevel(-1, axis=0)
+        uniques = pd.concat(uniques, names=by, copy=False)
+        uniques = uniques.droplevel(-1, axis=0)
     columns = df.columns
-    index_labels = df.index.names
     indicator = False
     if fill_value is not None and not explicit:
         # to get a name that does not exist in the columns
@@ -299,86 +244,63 @@ def _computations_complete(
     )
 
     if fill_value is not None:
-        if is_scalar(fill_value):
-            # faster when fillna operates on a Series basis
-            fill_value = {
-                col: fill_value for col in columns if out[col].hasnans
-            }
-        if explicit:
-            out = out.fillna(fill_value, downcast="infer")
-        else:
-            # keep only columns that are not part of column_checker
-            # IOW, we are excluding columns that were not used
-            # to generate the combinations
-            fill_value = {
-                col: value
-                for col, value in fill_value.items()
-                if col not in column_checker
-            }
-            if fill_value:
-                # when explicit is False
-                # use the indicator parameter to identify rows
-                # for `left_only`, and fill the relevant columns in fill_value
-                # with the associated value.
-                boolean_filter = out.loc[:, indicator] == "left_only"
-                out = out.drop(columns=indicator)
-                # iteration used here,
-                # instead of assign (which is also a for loop),
-                # to cater for scenarios where the column_name is not a string
-                # assign only works with keys that are strings
-                # Also, the output will be floats (for numeric types),
-                # even if all the columns could be integers
-                # user can always convert to int if required
-                for column_name, value in fill_value.items():
-                    # for categorical dtypes, set the categories first
-                    if isinstance(out[column_name].dtype, pd.CategoricalDtype):
-                        out[column_name] = out[column_name].cat.add_categories(
-                            [value]
-                        )
-                    out.loc[boolean_filter, column_name] = value
-
-    if index and (index_labels != out.index.names):
-        labels = [label for label in index_labels if label in out.index.names]
-        return out.reorder_levels(order=labels, axis="index")
+        null_columns = [col for col in columns if out[col].hasnans]
+        if null_columns:
+            if is_scalar(fill_value):
+                # faster when fillna operates on a Series basis
+                fill_value = {col: fill_value for col in null_columns}
+            if explicit:
+                out = out.fillna(fill_value)
+            else:
+                # keep only columns that are not part of column_checker
+                # IOW, we are excluding columns that were not used
+                # to generate the combinations
+                fill_value = {
+                    col: value
+                    for col, value in fill_value.items()
+                    if col not in column_checker
+                }
+                if fill_value:
+                    # when explicit is False
+                    # use the indicator parameter to identify rows
+                    # for `left_only`, and fill the relevant columns
+                    # in fill_value with the associated value.
+                    boolean_filter = out.loc[:, indicator] == "left_only"
+                    out = out.drop(columns=indicator)
+                    # iteration used here,
+                    # instead of assign (which is also a for loop),
+                    # to cater for scenarios where the column_name
+                    # is not a string
+                    # assign only works with keys that are strings
+                    # Also, the output wil be floats (for numeric types),
+                    # even if all the columns could be integers
+                    # user can always convert to int if required
+                    for column_name, value in fill_value.items():
+                        # for categorical dtypes, set the categories first
+                        if isinstance(
+                            out[column_name].dtype, pd.CategoricalDtype
+                        ):
+                            out[column_name] = out[
+                                column_name
+                            ].cat.add_categories([value])
+                        out.loc[boolean_filter, column_name] = value
     if not out.columns.equals(columns):
-        return out.reindex(columns=columns)
+        return out.loc(axis=1)[columns]
     return out
 
 
 def _generic_complete(
-    df: pd.DataFrame, columns: list, all_scalars: bool, index: bool, sort: bool
+    df: pd.DataFrame, columns: list, all_scalars: bool, sort: bool
 ):
     """Generate cartesian product for `_computations_complete`.
 
     Returns a DataFrame, with no duplicates.
     """
     if all_scalars:
-        if index:
-            uniques = {
-                col: pd.factorize(df.index.get_level_values(col), sort=sort)[
-                    -1
-                ]
-                for col in columns
-            }
-        else:
-            uniques = {
-                col: pd.factorize(df[col], sort=sort)[-1] for col in columns
-            }
+        uniques = {
+            col: pd.factorize(df[col], sort=sort)[-1] for col in columns
+        }
         uniques = _computations_expand_grid(uniques)
-        if index:
-            # it is assured that scalars cannot be single
-            # hence a MultiIndex
-            uniques = {key[0]: value for key, value in uniques.items()}
-            data = list(uniques.values())
-            uniques = pd.MultiIndex.from_arrays(data, names=uniques)
-            if df.columns.nlevels == 1:
-                columns = pd.Index([], name=df.columns.name)
-            else:
-                length = len(df.columns.names)
-                columns = pd.MultiIndex.from_arrays(
-                    [[]] * length, names=df.columns.names
-                )
-            return pd.DataFrame([], index=uniques, columns=columns, copy=False)
         uniques = pd.DataFrame(uniques, copy=False)
         uniques.columns = columns
         return uniques
@@ -387,42 +309,22 @@ def _generic_complete(
     for ind, column in enumerate(columns):
         if isinstance(column, dict):
             len_columns = len(columns)
-            column = _complete_column(column, df=df, index=index, sort=sort)
+            column = _complete_column(column, df=df, sort=sort)
             # iteration here avoids any potential index collision
             column = {
                 ind + len_columns + key: value for key, value in column.items()
             }
             uniques.update(column)
         else:
-            uniques[ind] = _complete_column(
-                column, df=df, index=index, sort=sort
-            )
+            uniques[ind] = _complete_column(column, df=df, sort=sort)
     uniques = _computations_expand_grid(uniques)
-    if index:
-        uniques = {key[-1]: value for key, value in uniques.items()}
-        if len(uniques) > 1:
-            data = list(uniques.values())
-            uniques = pd.MultiIndex.from_arrays(data, names=uniques)
-        else:
-            key = next(iter(uniques))
-            data = uniques[key]
-            uniques = pd.Index(data, name=key)
-        if df.columns.nlevels == 1:
-            columns = pd.Index([], name=df.columns.name)
-        else:
-            length = len(df.columns.names)
-            columns = pd.MultiIndex.from_arrays(
-                [[]] * length, names=df.columns.names
-            )
-        return pd.DataFrame([], index=uniques, columns=columns, copy=False)
-
     uniques = pd.DataFrame(uniques, copy=False)
     uniques.columns = uniques.columns.droplevel(0)
     return uniques
 
 
 @functools.singledispatch
-def _complete_column(column, df, index, sort):
+def _complete_column(column, df, sort):
     """
     Args:
         column: scalar/list/dict
@@ -434,15 +336,12 @@ def _complete_column(column, df, index, sort):
         or a dictionary of unique Pandas Series.
     """
 
-    if index:
-        _, arr = pd.factorize(df.index.get_level_values(column), sort=sort)
-    else:
-        _, arr = pd.factorize(df.loc(axis=1)[column], sort=sort)
+    _, arr = pd.factorize(df.loc(axis=1)[column], sort=sort)
     return pd.Series(arr, name=column)
 
 
 @_complete_column.register(list)  # noqa: F811
-def _sub_complete_column(column, df, index, sort):  # noqa: F811
+def _sub_complete_column(column, df, sort):  # noqa: F811
     """
     Args:
         column: list
@@ -452,41 +351,26 @@ def _sub_complete_column(column, df, index, sort):  # noqa: F811
     Returns:
         Pandas DataFrame
     """
+    outcome = df.loc(axis=1)[column]
 
-    if index:
-        outcome = df.index
-        exclude = [label for label in outcome.names if label not in column]
-        if exclude:
-            outcome = outcome.droplevel(level=exclude)
-        # ideally, there shouldn't be nulls in the index
-        exclude = [outcome.get_level_values(label).isna() for label in column]
-        exclude = np.column_stack(exclude).all(axis=1)
-        if exclude.any():
-            outcome = outcome[~exclude]
-        _, outcome = pd.factorize(outcome, sort=sort)
-        outcome.names = column
+    exclude = outcome.isna().all(axis=1)
 
-    else:
-        outcome = df.loc(axis=1)[column]
+    if exclude.any(axis=None):
+        outcome = outcome.loc[~exclude]
 
-        exclude = outcome.isna().all(axis=1)
+    exclude = outcome.duplicated()
 
-        if exclude.any(axis=None):
-            outcome = outcome.loc[~exclude]
+    if exclude.any():
+        outcome = outcome.loc[~exclude]
 
-        exclude = outcome.duplicated()
-
-        if exclude.any():
-            outcome = outcome.loc[~exclude]
-
-        if sort:
-            outcome = outcome.sort_values(by=column)
+    if sort:
+        outcome = outcome.sort_values(by=column)
 
     return outcome
 
 
 @_complete_column.register(dict)  # noqa: F811
-def _sub_complete_column(column, df, index, sort):  # noqa: F811
+def _sub_complete_column(column, df, sort):  # noqa: F811
     """
     Args:
         column: dictionary
@@ -499,10 +383,7 @@ def _sub_complete_column(column, df, index, sort):  # noqa: F811
 
     collection = {}
     for ind, (key, value) in enumerate(column.items()):
-        if index:
-            arr = apply_if_callable(value, df.index.get_level_values(key))
-        else:
-            arr = apply_if_callable(value, df[key])
+        arr = apply_if_callable(value, df)
         if not is_list_like(arr):
             raise ValueError(f"value for {key} should be a 1-D array.")
         if not hasattr(arr, "shape"):
@@ -552,30 +433,14 @@ def _data_checks_complete(
     Check is conducted to ensure that column names are not repeated.
     Also checks that the names in `columns` actually exist in `df`.
 
-    Returns `df`, `columns`, `column_checker`, `index`, `by`, `fill_value`,
+    Returns `df`, `columns`, `column_checker`, `by`, `fill_value`,
     and `explicit` if all checks pass.
     """
 
-    df_columns = df.columns
-    index_labels = (
-        df.index.names
-        if isinstance(df.index, pd.MultiIndex)
-        else [df.index.name]
-    )
     if by:
         if not isinstance(by, list):
             by = [by]
-        for label in by:
-            if label in df_columns:
-                continue
-            elif label in index_labels:
-                index = True
-                continue
-            else:
-                raise ValueError(
-                    f"{label} in by is neither in the dataframe's columns, "
-                    "nor is it a label in the dataframe's index names."
-                )
+        check_column(df, column_names=by, present=True)
 
     columns = [
         [*grouping] if isinstance(grouping, tuple) else grouping
@@ -592,38 +457,20 @@ def _data_checks_complete(
             column_checker.extend(grouping)
 
     # columns should not be duplicated across groups
+    # nor should it exist in `by`
     column_checker_no_duplicates = set()
     for column in column_checker:
         if column is None:
             raise ValueError("label in the columns argument cannot be None.")
         if column in column_checker_no_duplicates:
             raise ValueError(f"{column} should be in only one group.")
+        if by and (column in by):
+            raise ValueError(
+                f"{column} already exists as a label in the `by` argument."
+            )
         column_checker_no_duplicates.add(column)  # noqa: PD005
 
-    # columns should either be all in columns
-    # or all in index.names/index.name
-    # ideally there shouldn't be None in either index names or columns
-    index = False
-    for label in column_checker:
-        if by and (label in by):
-            raise ValueError(f"{label} already exists in by.")
-        if label in df_columns:
-            continue
-        elif label in index_labels:
-            index = True
-            continue
-        else:
-            raise ValueError(
-                f"{label} is neither in the dataframe's columns, "
-                "nor is it a label in the dataframe's index names."
-            )
-
-    if index:
-        for label in column_checker:
-            if label not in index_labels:
-                raise ValueError(
-                    f"{label} not found in the dataframe's index names."
-                )
+    check_column(df, column_names=column_checker, present=True)
 
     check("explicit", explicit, [bool])
 
@@ -644,4 +491,4 @@ def _data_checks_complete(
                     f"The value for {column_name} should be a scalar."
                 )
 
-    return columns, column_checker, index, sort, by, fill_value, explicit
+    return columns, column_checker, sort, by, fill_value, explicit